test_seopendata_csv.rb - rename/refactor as test_seopendata_csv_schema_converter.rb
This tests conversions, rather than schemas per-se. As such it
describes the same namespace, but specifically the part where
converters are created using schemas.

Add a couple of tests which ensure that spurious field IDs are
reported when added to the input or outputs of the conversion blocks.

The tests all pass.