se_open_data/utils/deployment.rb - performs rsync deployment
Isolates the rsync messiness in one place, and avoid needing to tweak
rsync work-arounds and hacks in multiple places.

Specifically, I want to allow deployments to the local filesystem,
when run as a cron-job on the develppment server.  This makes the
deployment simpler, and more secure (no need to deploy ssh keys etc.)

Also, I want to fix the bug in which files are deployed with crazy
ownership (typically, user ubuntu, group lxd, but this presumably
depends on the deploying user's set up).  Ownership should be settable
to something like 'www-data'.

And if possible, avoid using rsync *and* ssh.  Use rsync in a way
which should work on rsync servers, and especially with rsync to
accounts with log-ins constrained to allow only rsync in a specific
way with a specific key.  Because we will probably want to enforce
that kind of security at some point in the future.

This is still ugly. Perhaps we should use capistrano or something
else. But I hope it will do.