document-geocoder.rb - don't expect non-supplied option
In practice the Makefile doesn't supply the api_key option, and the
fall-back is hardwird to geoapifyAPI.txt, so why not hardwire the
password label to geoapifyAPI.txt too?  At least for now, to get
things working.

Later this script might need to be refactored into the library code
anyway.