se_open_data/initiative/rdf/config.rb - ensure uri_prefix ends with slash
This is kind of hacky, but we need to ensure that links to the
uri_prefix link to the index page. If they don't end with a slash,
this is ambiguous, since it could be the URI of a term, and the
lod.coop redirectory will add a .html suffix when rewriting it,
instead of an /index.html suffix.