se_open_data.gemspec, Gemfile - allow this repo to be installed as a Gem
We want to be able to install this from the git repo as a Gem, using a
Gemfile dependency like this:

    gem "se_open_data", :github => "SolidarityEconomyAssociation/se-open-data", :branch => "0.1.0"

To allow the dependencies of se_open_data to be installed transitively
from that, this repo needs to hace a .gemspec file added which defines
the dependencies, and a Gemfile which delegates to the gemspec. This
is what we add here.